# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: family-finances API
  version: 0.1.0
  description: >-
    The JSON HTTP API of the family-finances backend. This document is
    hand-written and is the source of truth for the API surface (see
    openapi/README.md). Operations tagged `browser-flow` and marked
    `x-internal: true` are browser-navigation redirects, not part of the
    generated typed client.
servers:
  - url: /
    description: Same-origin — the Go backend serves both the API and the SPA.
...
paths:
...
  /api/accounts/{id}/balance:
    get:
      operationId: getAccountBalance
      summary: The account's live balance
      description: >-
        Always computed live, never cached — see the account-entries
        capability. Equals the latest non-deleted balance_adjustment at or
        before as_of (or 0 if none), plus every non-deleted transaction
        after it, up to as_of.
      tags: [entries]
      parameters:
        - $ref: "#/components/parameters/AccountId"
        - name: as_of
          in: query
          required: false
          description: RFC3339 timestamp. Defaults to now.
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: The computed balance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Balance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/categories:
    get:
      operationId: getCategories
      summary: List the full category tree
      tags: [categories]
      responses:
        "200":
          description: Every category, including its parent_id.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: postCategories
      summary: Create a category (admin only)
      tags: [categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryWrite"
      responses:
        "201":
          description: The created category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/categories/{id}:
    patch:
      operationId: patchCategory
      summary: Update a category (admin only)
      description: >-
        Reparenting onto the category itself or one of its own descendants
        is rejected with 422.
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryWrite"
      responses:
        "200":
          description: The updated category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    delete:
      operationId: deleteCategory
      summary: Delete a category (admin only)
      description: >-
        409 if it has child categories or is referenced by a non-deleted
        entry — deletion is never cascaded.
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: The category is deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/entries:
    get:
      operationId: getEntries
      summary: List the caller's entries, filtered/searched/sorted, cursor-paginated
      tags: [entries]
      parameters:
        - name: account_id
          in: query
          required: false
          description: Repeatable. Omitted means every account the caller owns.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: category_id
          in: query
          required: false
          description: Matches this category and every descendant.
          schema:
            type: string
        - name: tag_id
          in: query
          required: false
          schema:
            type: string
        - name: kind
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EntryKind"
        - name: from
          in: query
          required: false
          description: Inclusive booking_timestamp lower bound, RFC3339.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Inclusive booking_timestamp upper bound, RFC3339.
          schema:
            type: string
            format: date-time
        - name: q
          in: query
          required: false
          description: Case-insensitive substring match against title or description.
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [booking_timestamp, amount]
            default: booking_timestamp
        - name: dir
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: after
          in: query
          required: false
          description: Opaque cursor from a previous response's next_cursor.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: A page of matching entries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: postEntries
      summary: Create an entry
      description: >-
        422 when account_id does not name an account the caller owns, or
        names a disabled account.
      tags: [entries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryCreate"
      responses:
        "201":
          description: The created entry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/entries/{id}:
    get:
      operationId: getEntry
      summary: Get an entry
      tags: [entries]
      parameters:
        - $ref: "#/components/parameters/EntryId"
      responses:
        "200":
          description: The entry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchEntry
      summary: Update an entry
      description: >-
        account_id and kind are immutable — there is no field for them on
        this request body at all.
      tags: [entries]
      parameters:
        - $ref: "#/components/parameters/EntryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryUpdate"
      responses:
        "200":
          description: The updated entry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteEntry
      summary: Soft-delete an entry
      description: One-way — there is no undelete endpoint.
      tags: [entries]
      parameters:
        - $ref: "#/components/parameters/EntryId"
      responses:
        "204":
          description: The entry is soft-deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ff_session
    bearerToken:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: No valid session was presented.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The authenticated user is not an admin.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The request conflicts with the resource's current state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: >-
        No such resource — including one that belongs to a different owner,
        which behaves identically to nonexistent (existence is not
        disclosed across owners).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: A field in the request failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: >-
        The request is well-formed but violates a business rule (e.g. a
        reparent that would create a category cycle, or an entry against a
        disabled account).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  parameters:
    AccountId:
      name: id
      in: path
      required: true
      schema:
        type: string
    EntryId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        request_id:
          type: string
...
    Category:
      type: object
      required: [id, name, created_at]
      properties:
        id:
          type: string
        parent_id:
          type: string
          description: Absent for a root category.
        name:
          type: string
        created_at:
          type: string
          format: date-time
    CategoryWrite:
      type: object
      properties:
        name:
          type: string
        parent_id:
          type: string
          nullable: true
          description: Explicit null makes it a root category.
...
    EntryKind:
      type: string
      enum: [transaction, balance_adjustment]
    Entry:
      type: object
      required:
        - id
        - account_id
        - kind
        - amount
        - booking_timestamp
        - title
        - tag_ids
        - created_at
        - updated_at
      properties:
        id:
          type: string
        account_id:
          type: string
        kind:
          $ref: "#/components/schemas/EntryKind"
        amount:
          type: integer
          format: int64
          description: >-
            Integer minor units at a fixed 4 decimal places (e.g. 105000
            represents 10.5000 in the account's currency). Not
            configurable — see account-entries.
        booking_timestamp:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: string
        category_id:
          type: string
          description: >-
            Required for a transaction, optional for a balance_adjustment.
        tag_ids:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EntryCreate:
      type: object
      required: [account_id, kind, amount, booking_timestamp, title]
      properties:
        account_id:
          type: string
        kind:
          $ref: "#/components/schemas/EntryKind"
        amount:
          type: integer
          format: int64
        booking_timestamp:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: string
        category_id:
          type: string
        tag_ids:
          type: array
          items:
            type: string
    EntryUpdate:
      type: object
      description: >-
        No account_id or kind field — both are immutable after creation.
      properties:
        amount:
          type: integer
          format: int64
        booking_timestamp:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: string
        category_id:
          type: string
          nullable: true
          description: >-
            Explicit null clears it (only valid when the entry's kind is
            balance_adjustment).
        tag_ids:
          type: array
          items:
            type: string
          description: Replaces the full set, including clearing it with [].
    EntryPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
        next_cursor:
          type: string
          nullable: true
    Balance:
      type: object
      required: [balance]
      properties:
        balance:
          type: integer
          format: int64
security:
  - sessionCookie: []
  - bearerToken: []